requirements.txt
Adafruit-Blinka==8.6.0
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-dht==3.7.7
adafruit-circuitpython-requests==1.12.9
adafruit-circuitpython-typing==1.8.2
Adafruit-PlatformDetect==3.32.0
Adafruit-PureIO==1.1.9
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
idna==3.4
pycryptodome==3.15.0
pyftdi==0.54.0
pyserial==3.5
python-dotenv==0.21.0
pyusb==1.2.1
requests==2.28.1
tuya-connector-python==0.1.2
typing_extensions==4.4.0
urllib3==1.26.12
websocket-client==1.4.1
